+++ packages/components/src/Dialog/Layout/DialogContent.test.tsx
@@ -26,17 +26,34 @@
 
 import 'jest-styled-components'
 import React from 'react'
-import { assertSnapshot } from '@looker/components-test-utils'
+import { assertSnapshot, renderWithTheme } from '@looker/components-test-utils'
 import { DialogContent } from './DialogContent'
 
-test('DialogContent - no overflow', () => {
-  assertSnapshot(<DialogContent>Stuff</DialogContent>)
-})
+describe('DialogContent', () => {
+  test('Snapshot', () => {
+    assertSnapshot(<DialogContent>Stuff</DialogContent>)
+  })
 
-test('DialogContent - no overflow', () => {
-  assertSnapshot(
-    <DialogContent>
-      <div style={{ height: '4rem' }}>Stuff</div>
-    </DialogContent>
-  )
+  test('Snapshot - No overflow', () => {
+    assertSnapshot(
+      <DialogContent>
+        <div style={{ height: '4rem' }}>Stuff</div>
+      </DialogContent>
+    )
+  })
+
+  test('borderBottom', () => {
+    const { getByText } = renderWithTheme(
+      <>
+        <DialogContent>No border</DialogContent>
+        <DialogContent borderBottom>Has border</DialogContent>
+      </>
+    )
+
+    const noBorder = getByText('No border')
+    expect(noBorder.parentNode).toHaveStyle('border-bottom: none')
+
+    const hasBorder = getByText('Has border')
+    expect(hasBorder.parentNode).toHaveStyle('border-bottom: 1px solid #DEE1E5')
+  })
 })