+++ storybook/src/Overlays/Popovers/Testing.stories.tsx
@@ -30,7 +30,7 @@ import {
   Button,
   ButtonOutline,
   ButtonTransparent,
-  DialogManager,
+  Dialog,
   FieldSelect,
   Heading,
   Menu,
@@ -41,6 +41,7 @@ import {
   Paragraph,
   Popover,
   PopoverContent,
+  SpaceVertical,
   usePopover,
   useToggle,
 } from '@looker/components'
@@ -85,7 +86,7 @@ export const PopoverFocusTrap = () => {
         content={
           <PopoverContent p="large" width="360px">
             <Paragraph>
-              Does tabbing focus only loop through these 3 buttons?
+              Does tabbing focus only loop through these 3 buttons &amp; Select?
             </Paragraph>
             <Paragraph>
               Does clicking (or mousedown) each trigger an alert?
@@ -110,6 +111,9 @@ export const PopoverFocusTrap = () => {
               aria-label="Fruits"
               defaultValue="1"
             />
+            <Paragraph>
+              Does it scroll here when the Select is closed?
+            </Paragraph>
             <Paragraph>Long text</Paragraph>
             <Paragraph>Long text</Paragraph>
             <Paragraph>Long text</Paragraph>
@@ -128,37 +132,84 @@ export const PopoverFocusTrap = () => {
       >
         <Button mt="medium">Open Focus Trap Test Popover</Button>
       </Popover>
+      <Paragraph>Does it scroll here when the Popover is closed?</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
+      <Paragraph>Long text</Paragraph>
     </Box>
   )
 }
 
-export const MenuOpenDialog = () => {
+export const OverlayOpenDialog = () => {
+  const { value, setOn, setOff } = useToggle()
   function openAlert() {
     alert(`It's working!`)
   }
   return (
-    <Box mt="large">
-      <Heading>Menu Opening Dialog</Heading>
+    <SpaceVertical mt="large" align="start">
+      <Heading>Popover Opening a Dialog</Heading>
+      <Popover
+        content={
+          <Button m="large" onClick={setOn}>
+            Open Dialog
+          </Button>
+        }
+      >
+        <Button>Open Popover</Button>
+      </Popover>
+      <Dialog isOpen={value} onClose={setOff}>
+        <DialogContent>
+          <SpaceVertical align="start">
+            <Paragraph>Try opening the Select and picking an option:</Paragraph>
+            <FieldSelect
+              label="Default Value"
+              width={300}
+              options={options}
+              aria-label="Fruits"
+              defaultValue="1"
+            />
+            <Paragraph>Try clicking the button:</Paragraph>
+            <Button onClick={openAlert}>Open Alert</Button>
+          </SpaceVertical>
+        </DialogContent>
+      </Dialog>
+      <Heading>Menu Opening a Dialog</Heading>
       <Menu>
         <MenuDisclosure tooltip="Select your favorite kind">
           <Button mr="small" mt="medium">
             Open Menu
           </Button>
         </MenuDisclosure>
-        <DialogManager
-          content={
-            <DialogContent>
-              <Paragraph>Some content inside the Dialog</Paragraph>
-              <Button onClick={openAlert}>Open Alert</Button>
-            </DialogContent>
-          }
-        >
-          <MenuList>
-            <MenuItem>Open Dialog</MenuItem>
-          </MenuList>
-        </DialogManager>
+        <MenuList>
+          <MenuItem onClick={setOn}>Open Dialog</MenuItem>
+        </MenuList>
       </Menu>
-    </Box>
+    </SpaceVertical>
   )
 }
 
@@ -177,7 +228,7 @@ export const RenderProps = () => (
 
 export const RenderPropsSpread = () => (
   <Popover content={popoverContent}>
-    {(...props) => <button {...props}>Test</button>}
+    {(props) => <button {...props}>Test</button>}
   </Popover>
 )